# Optimize Integer decoding from bytes