@asterite 2022-02-03
There's actually no memory allocation involved here. This works slightly faster in this PR because some bound checks are removed, which makes this easily prone to segfault. I'm not sure we should do that.

I'm not on the computer, but if I have time I will send code that makes this change segfault.