@straight-shoota 2022-02-03
Yes, the change to `.encode` is definitely incorrect. It removes bounds checking on `bytes` and doesn't validate that it's not `read_only`.

I believe the change to `.decode` should be safe, though. It still uses `Slice#copy_to` which includes bounds checking. And we can be certain of the buffer size, so the pointer parameter should not be an issue.