@carlhoerberg 2022-02-03
Removed `encode`, it was an afterthought and didn't improve performance anyway.