@jgaskins 2022-03-22
@carlhoerberg Your benchmark results seem to have run into the issue where each iteration is so fast that the results are inaccurate. `Benchmark.ips` can't be trusted when the results are in the 1-3ns range.

I updated the benchmark code to run more iterations inside the `report` block and the only notable change in performance was on `Int128`. Everything else was within ±1% and even closer on ARM.

<details><summary>Updated code</summary>

```crystal
require "benchmark"

module IO::ByteFormat
  {% for mod in %w(LittleEndian BigEndian) %}
    module {{mod.id}}
      {% for type, i in %w(Int8 UInt8 Int16 UInt16 Int32 UInt32 Int64 UInt64 Int128 UInt128) %}
        {% bytesize = 2 ** (i // 2) %}
        def self.decode2(type : {{type.id}}.class, bytes : Bytes)
          buffer = uninitialized UInt8[{{bytesize}}]
          bytes.copy_to(buffer.to_unsafe, {{bytesize}})
          buffer.reverse! unless SystemEndian == self
          buffer.unsafe_as({{type.id}})
        end
      {% end %}
    end
  {% end %}
end

b = Bytes.new(128) { 1u8 }
iterations = 1_000

{% for mod in %w(LittleEndian BigEndian) %}
  {% for type, i in %w(Int8 UInt8 Int16 UInt16 Int32 UInt32 Int64 UInt64 Int128 UInt128) %}
    Benchmark.ips do |x|
      x.report("old decode {{mod.id}} {{ type.id }}") do
        iterations.times { IO::ByteFormat::{{mod.id}}.decode({{type.id}}, b) }
      end
      x.report("new decode {{mod.id}} {{ type.id }}") do
        iterations.times { IO::ByteFormat::{{mod.id}}.decode2({{type.id}}, b) }
      end
    end
  {% end %}
{% end %}
```

</details>

<details><summary>Diff from the original benchmark code</summary>

```diff
19a20
> iterations = 1_000
25c26
<         IO::ByteFormat::{{mod.id}}.decode({{type.id}}, b)
---
>         iterations.times { IO::ByteFormat::{{mod.id}}.decode({{type.id}}, b) }
28c29
<         IO::ByteFormat::{{mod.id}}.decode2({{type.id}}, b)
---
>         iterations.times { IO::ByteFormat::{{mod.id}}.decode2({{type.id}}, b) }
```

</details>

<details><summary>Intel results</summary>

```
old decode LittleEndian Int8   4.44M (224.98ns) (± 1.25%)  0.0B/op   1.01× slower
new decode LittleEndian Int8   4.49M (222.73ns) (± 0.78%)  0.0B/op        fastest
old decode LittleEndian UInt8   4.48M (223.25ns) (± 0.94%)  0.0B/op        fastest
new decode LittleEndian UInt8   4.44M (225.25ns) (± 1.85%)  0.0B/op   1.01× slower
old decode LittleEndian Int16   4.49M (222.81ns) (± 1.23%)  0.0B/op        fastest
new decode LittleEndian Int16   4.45M (224.68ns) (± 0.75%)  0.0B/op   1.01× slower
old decode LittleEndian UInt16   4.46M (224.09ns) (± 1.10%)  0.0B/op   1.01× slower
new decode LittleEndian UInt16   4.49M (222.90ns) (± 1.23%)  0.0B/op        fastest
old decode LittleEndian Int32   4.50M (222.37ns) (± 1.15%)  0.0B/op        fastest
new decode LittleEndian Int32   4.48M (223.13ns) (± 0.88%)  0.0B/op   1.00× slower
old decode LittleEndian UInt32   4.44M (225.38ns) (± 0.66%)  0.0B/op   1.01× slower
new decode LittleEndian UInt32   4.47M (223.47ns) (± 1.75%)  0.0B/op        fastest
old decode LittleEndian Int64   4.44M (225.45ns) (± 2.39%)  0.0B/op        fastest
new decode LittleEndian Int64   4.43M (225.85ns) (± 1.15%)  0.0B/op   1.00× slower
old decode LittleEndian UInt64   4.46M (224.41ns) (± 1.31%)  0.0B/op   1.00× slower
new decode LittleEndian UInt64   4.48M (223.41ns) (± 1.65%)  0.0B/op        fastest
old decode LittleEndian Int128   4.49M (222.78ns) (± 1.03%)  0.0B/op        fastest
new decode LittleEndian Int128   4.46M (224.29ns) (± 1.27%)  0.0B/op   1.01× slower
old decode LittleEndian UInt128   4.48M (223.31ns) (± 0.77%)  0.0B/op   1.00× slower
new decode LittleEndian UInt128   4.49M (222.76ns) (± 0.72%)  0.0B/op        fastest
old decode BigEndian Int8   4.48M (223.07ns) (± 0.94%)  0.0B/op        fastest
new decode BigEndian Int8   4.45M (224.97ns) (± 0.67%)  0.0B/op   1.01× slower
old decode BigEndian UInt8   4.48M (223.39ns) (± 0.97%)  0.0B/op   1.01× slower
new decode BigEndian UInt8   4.53M (220.96ns) (± 0.73%)  0.0B/op        fastest
old decode BigEndian Int16   4.52M (221.31ns) (± 0.82%)  0.0B/op        fastest
new decode BigEndian Int16   4.48M (223.06ns) (± 0.93%)  0.0B/op   1.01× slower
old decode BigEndian UInt16   4.47M (223.78ns) (± 1.01%)  0.0B/op   1.01× slower
new decode BigEndian UInt16   4.51M (221.84ns) (± 0.67%)  0.0B/op        fastest
old decode BigEndian Int32   4.50M (222.36ns) (± 0.93%)  0.0B/op        fastest
new decode BigEndian Int32   4.48M (223.10ns) (± 0.83%)  0.0B/op   1.00× slower
old decode BigEndian UInt32   4.45M (224.72ns) (± 0.88%)  0.0B/op   1.01× slower
new decode BigEndian UInt32   4.50M (222.18ns) (± 1.17%)  0.0B/op        fastest
old decode BigEndian Int64   4.50M (221.98ns) (± 1.16%)  0.0B/op        fastest
new decode BigEndian Int64   4.48M (223.24ns) (± 0.84%)  0.0B/op   1.01× slower
old decode BigEndian UInt64   4.49M (222.73ns) (± 0.72%)  0.0B/op   1.00× slower
new decode BigEndian UInt64   4.50M (222.26ns) (± 0.65%)  0.0B/op        fastest
old decode BigEndian Int128 154.46k (  6.47µs) (± 1.14%)  0.0B/op   1.43× slower
new decode BigEndian Int128 220.18k (  4.54µs) (± 0.74%)  0.0B/op        fastest
old decode BigEndian UInt128 154.52k (  6.47µs) (± 1.05%)  0.0B/op   1.42× slower
new decode BigEndian UInt128 219.74k (  4.55µs) (± 1.08%)  0.0B/op        fastest
```

</details>

<details><summary>ARM64 results</summary>

```
old decode LittleEndian Int8   3.09M (323.65ns) (± 0.33%)  0.0B/op        fastest
new decode LittleEndian Int8   3.09M (323.79ns) (± 0.15%)  0.0B/op   1.00× slower
old decode LittleEndian UInt8   3.09M (323.69ns) (± 0.32%)  0.0B/op        fastest
new decode LittleEndian UInt8   3.09M (324.13ns) (± 0.47%)  0.0B/op   1.00× slower
old decode LittleEndian Int16   3.09M (323.54ns) (± 0.21%)  0.0B/op        fastest
new decode LittleEndian Int16   3.09M (323.73ns) (± 0.15%)  0.0B/op   1.00× slower
old decode LittleEndian UInt16   3.09M (323.68ns) (± 0.40%)  0.0B/op        fastest
new decode LittleEndian UInt16   3.09M (323.77ns) (± 0.15%)  0.0B/op   1.00× slower
old decode LittleEndian Int32   3.09M (324.00ns) (± 0.89%)  0.0B/op   1.00× slower
new decode LittleEndian Int32   3.09M (323.90ns) (± 0.26%)  0.0B/op        fastest
old decode LittleEndian UInt32   3.09M (324.02ns) (± 0.86%)  0.0B/op   1.00× slower
new decode LittleEndian UInt32   3.09M (323.86ns) (± 0.20%)  0.0B/op        fastest
old decode LittleEndian Int64   3.09M (323.46ns) (± 0.14%)  0.0B/op        fastest
new decode LittleEndian Int64   3.09M (323.91ns) (± 0.31%)  0.0B/op   1.00× slower
old decode LittleEndian UInt64   3.09M (323.59ns) (± 0.23%)  0.0B/op        fastest
new decode LittleEndian UInt64   3.09M (323.93ns) (± 0.34%)  0.0B/op   1.00× slower
old decode LittleEndian Int128   3.09M (323.47ns) (± 0.15%)  0.0B/op        fastest
new decode LittleEndian Int128   3.09M (323.92ns) (± 0.24%)  0.0B/op   1.00× slower
old decode LittleEndian UInt128   3.09M (323.49ns) (± 0.14%)  0.0B/op        fastest
new decode LittleEndian UInt128   3.09M (323.94ns) (± 0.39%)  0.0B/op   1.00× slower
old decode BigEndian Int8   3.09M (323.54ns) (± 0.20%)  0.0B/op        fastest
new decode BigEndian Int8   3.09M (323.83ns) (± 0.27%)  0.0B/op   1.00× slower
old decode BigEndian UInt8   3.09M (323.43ns) (± 0.12%)  0.0B/op        fastest
new decode BigEndian UInt8   3.09M (323.77ns) (± 0.15%)  0.0B/op   1.00× slower
old decode BigEndian Int16   3.09M (323.46ns) (± 0.13%)  0.0B/op        fastest
new decode BigEndian Int16   3.08M (324.20ns) (± 0.86%)  0.0B/op   1.00× slower
old decode BigEndian UInt16   3.09M (323.62ns) (± 0.26%)  0.0B/op        fastest
new decode BigEndian UInt16   3.09M (323.82ns) (± 0.16%)  0.0B/op   1.00× slower
old decode BigEndian Int32   3.09M (323.47ns) (± 0.16%)  0.0B/op        fastest
new decode BigEndian Int32   3.09M (323.88ns) (± 0.22%)  0.0B/op   1.00× slower
old decode BigEndian UInt32   3.09M (323.45ns) (± 0.15%)  0.0B/op        fastest
new decode BigEndian UInt32   3.09M (323.78ns) (± 0.16%)  0.0B/op   1.00× slower
old decode BigEndian Int64   3.09M (323.49ns) (± 0.18%)  0.0B/op        fastest
new decode BigEndian Int64   3.09M (323.77ns) (± 0.15%)  0.0B/op   1.00× slower
old decode BigEndian UInt64   3.09M (323.65ns) (± 0.37%)  0.0B/op        fastest
new decode BigEndian UInt64   3.09M (323.87ns) (± 0.22%)  0.0B/op   1.00× slower
old decode BigEndian Int128 159.81k (  6.26µs) (± 0.32%)  0.0B/op   1.66× slower
new decode BigEndian Int128 265.72k (  3.76µs) (± 0.14%)  0.0B/op        fastest
old decode BigEndian UInt128 159.77k (  6.26µs) (± 0.30%)  0.0B/op   1.66× slower
new decode BigEndian UInt128 265.63k (  3.76µs) (± 0.18%)  0.0B/op        fastest
```

</details>

Seems like LLVM might be optimizing the stack allocation out of the resulting binary. I'm surprised your results were so consistently one-sided.